pyproject.toml
[project]
name = "fintech-fraud-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "kafka-python>=2.3.0",
    "lightgbm>=4.6.0",
    "mlflow>=3.8.0",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
]
